[dev] Fixed a cast error from size_t to unsigned int

+++ code/mesh/src/mesh_manipulations.cpp
@@ -161,7 +161,7 @@ void write_binary_stl(const VectorOfVectorOfPoints& stl, std::ostream& os)
     //const std::string header(80, ' ');
     //os.write(header.c_str(), 80);
     for (size_t i = 0 ; i < 80 ; ++i) os.put(1);
-    const unsigned int nFaces = stl.size();
+    const unsigned int nFaces = (unsigned int)stl.size();
     os.write(reinterpret_cast<const char*>(&nFaces), sizeof(nFaces));
 
     typedef unsigned short uint16;
